requirements.txt
PySide6==6.8.2.1
Pillow==11.1.0
pyinstaller==6.12.0